+++ enclone_args/src/lib.rs
@@ -1,6 +1,7 @@
 // Copyright (c) 2021 10x Genomics, Inc. All rights reserved.
 
 pub mod load_gex;
+pub mod load_gex_core;
 pub mod proc_args;
 pub mod proc_args2;
 pub mod proc_args3;
